Update db models

+++ backend/database/models/menu_field_product.py
@@ -8,7 +8,7 @@ class MenuFieldProduct(Model):
     """
 
     id = fields.IntField(pk=True)
-    price = fields.FloatField()
+    price = fields.DecimalField(max_digits=10, decimal_places=2)
     product = fields.ForeignKeyField(
         model_name="models.Product",
         on_delete=fields.RESTRICT,
